+++ backend/inpatient/views.py
@@ -1,6 +1,6 @@
 import logging
 from io import BytesIO
-
+from rest_framework.views import APIView
 from django.shortcuts import get_object_or_404
 from django.db.models import Prefetch
 from django_filters.rest_framework import DjangoFilterBackend
@@ -20,12 +20,13 @@ from company.models import Company
 from laboratory.models import LabTestRequest, PatientSample
 from patient.models import Triage, AttendanceProcess, PrescribedDrug
 
+from .utils import (generate_discharge_summary_pdf)
 from .filters import WardFilter, PatientAdmissionFilter
-from .models import (Bed, PatientAdmission, PatientDischarge, Ward,
-                     WardNurseAssignment)
+from .models import (Bed, PatientAdmission, PatientDischarge, Ward, WardNurseAssignment)
 from .serializers import (BedSerializer, PatientAdmissionSerializer,
-                          PatientDischargeSerializer,
-                          WardNurseAssignmentSerializer, WardSerializer)
+                        PatientDischargeSerializer,
+                        WardNurseAssignmentSerializer, WardSerializer)
+
 
 
 logger = logging.getLogger(__name__)
@@ -170,113 +171,11 @@ class PatientDischargeViewset(viewsets.ModelViewSet):
             return Response({"error": f"Failed to discharge patient: {str(e)}"}, status=status.HTTP_400_BAD_REQUEST)
 
 
-def download_discharge_summary(request, discharge_id):
-    """
-    Generate a discharge summary PDF for the specified discharge, including patient details,
-    last lab results, prescription, vitals, and type-specific data (referral or cause of death).
-    """
-    
-    try:
-        discharge = get_object_or_404(PatientDischarge, pk=discharge_id)
-        admission = discharge.admission
-        patient = admission.patient
-        company = Company.objects.first()
-
-        # Last lab result
-        last_lab_request = LabTestRequest.objects.filter(
-            process__attendanceprocess__patient=patient
-        ).order_by('-requested_on').first()
-        lab_results = []
-        if last_lab_request:
-            samples = PatientSample.objects.filter(lab_test_request=last_lab_request)
-            lab_results = [{
-                'test_name': last_lab_request.test_profile.name,
-                'result': sample.result or 'N/A',
-                'specimen': sample.specimen.name,
-                'date': sample.created_on
-            } for sample in samples]
-
-        # Last prescription
-        last_attendance = AttendanceProcess.objects.filter(
-            patient=patient
-        ).order_by('-created_at').first()
-        prescription_details = []
-        if last_attendance and last_attendance.prescription:
-            prescribed_drugs = PrescribedDrug.objects.filter(prescription=last_attendance.prescription)
-            prescription_details = [{
-                'medication': drug.item.name,
-                'dosage': drug.dosage,
-                'frequency': drug.frequency,
-                'duration': drug.duration,
-                'prescribed_by': last_attendance.prescription.created_by.get_fullname() if last_attendance.prescription.created_by else 'N/A',
-                'date': last_attendance.prescription.date_created
-            } for drug in prescribed_drugs]
-
-        # Latest triage vitals
-        last_triage = Triage.objects.filter(
-            attendanceprocess__patient=patient
-        ).order_by('-date_created').first()
-        triage_data = {}
-        if last_triage:
-            triage_data = {
-                'blood_pressure': f"{last_triage.systolic}/{last_triage.diastolic}" if last_triage.systolic and last_triage.diastolic else 'N/A',
-                'pulse': last_triage.pulse,
-                'temperature': last_triage.temperature,
-                'recorded_at': last_triage.date_created
-            }
-
-        company_logo_url = request.build_absolute_uri(company.logo.url) if company.logo else None
-
-        # Prepare context
-        context = {
-            'company': {
-            'name': company.name,
-            'address1': company.address1,
-            'phone1': company.phone1,
-            'email1': company.email1,
-            'logo_url': company_logo_url
-            },
-            'patient': {
-                'name': f"{patient.first_name} {patient.second_name}",
-                'age': patient.age,
-                'gender': patient.get_gender_display(),
-                'unique_id': patient.unique_id
-            },
-            'admission': {
-                'id': admission.admission_id,
-                'reason': admission.reason_for_admission,
-                'admitted_at': admission.admitted_at,
-                'ward': admission.ward.name if admission.ward else 'N/A',
-                'bed': admission.bed.bed_number if admission.bed else 'N/A'
-            },
-            'discharge': {
-                'type': discharge.get_discharge_types_display(),
-                'notes': discharge.discharge_notes or 'N/A',
-                'discharged_at': discharge.discharged_at,
-                'discharged_by': discharge.discharged_by.get_fullname() if discharge.discharged_by else 'N/A'
-            },
-            'lab_results': lab_results,
-            'prescription': prescription_details,
-            'triage': triage_data
-        }
-
-        # Type-specific data
-        if discharge.discharge_types == 'referral' and discharge.referral:
-            context['referral'] = {
-                'service': discharge.referral.get_service_display(),
-                'note': discharge.referral.note,
-                'email': discharge.referral.email,
-                'referred_by': discharge.referral.referred_by.get_fullname() if discharge.referral.referred_by else 'N/A'
-            }
-       
-
-        html_template = get_template('discharge_summary.html').render(context)
-        pdf_file = HTML(string=html_template).write_pdf()
+class DownloadDischargeSummaryView(APIView):
+    def get(self, request, admission_id):
+        pdf_file, error = generate_discharge_summary_pdf(admission_id, request)
+        if error:
+            return Response(error, status=404)
         response = HttpResponse(pdf_file, content_type='application/pdf')
-        response['Content-Disposition'] = f'attachment; filename="discharge_summary_{admission.admission_id}.pdf"'
+        response['Content-Disposition'] = f'attachment; filename=\"discharge_summary_{admission_id}.pdf\"'
         return response
-
-    except Exception as e:
-        logger.error(f"Failed to generate discharge summary: {str(e)}", exc_info=True)
-        return HttpResponse(f"Failed to generate discharge summary: {str(e)}", status=500)
-
