+++ backend/laboratory/signals.py
@@ -1,7 +1,7 @@
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 from .models import LabTestRequestPanel
-from easymed.celery_tasks import deduct_test_kit
+from laboratory.tasks import deduct_test_kit
 
 @receiver(post_save, sender=LabTestRequestPanel)
 def trigger_test_kit_deduction(sender, instance, **kwargs):
